+++ moredocs/Releases.md
@@ -1,6 +1,12 @@
 # @goodware/log Release History
 
-## 3.4.0 2020-07-05
+## 3.4.2 2020-07-31
+
+- Add new meta key 'stack' because it is supported by the data-enabled console
+- `loggers.info(undefined, {error: new Error('x')})` only outputs one message to the console
+- Remove feature that replaces Error instances with text in data because it was redundant and noisy
+
+## 3.4.0-3.4.1 2020-07-05
 
 ### Features
 
@@ -9,6 +15,7 @@
 ### Bug fixes
 
 - [#41](https://github.com/good-ware/js-log/issues/41) AWS logstream name output when restart() is called
+- Fix TaskLogger when the function throws an exception
 
 ## 3.3.6-3.3.8 2020-06-19
 
@@ -176,3 +183,15 @@ Bug fixes
 
 1. Allow message in `logger.levelName([tags], message)` to be an object with message, tags, and context keys
 2. Bug fix: 'tags' value was lost in loggers.child(tags).logger('loggerName')
+
+## 1.3.5 2021-03-23
+
+- The Ready and Stopping messages are logged at the default level
+
+## 1.3.3 2021-03-22
+
+- Add logging level methods to Loggers (e.g., error) (API change)
+- Documentation (hide more with @ignore)
+- Strip @goodware/log functions from stack traces
+- Support Node 8 and higher by changing process.off() to process.removeListener()
+- Link README to RunKit example